requirements.txt
Flask
flask-sqlalchemy
flask-migrate
python-dotenv
pytest
pylint
pep8
autopep8